Remove incorrect initialization of GCLK2
This code claims to configure GCLK2 to use OSC8M, however it doesn't do
this.  It uses the unshifted field GCLK_GENCTRL_SRC_OSC8M_Val when it
should have used GCLK_GENCTRL_SRC_OSC8M, which has been shifted by 8
bits.

Because of this, it's actually enabling GCLK6 and setting it to use XOSC
as its input.  Which is just totally wrong.

As GCLK2 is not used by anything in the bootloader, nothing breaks.

There is no reason to enable GCLK2, nor GCLK6, so just remove this code.

This saves 24 bytes of code.

This bug was created in commit 33cf8e206a5f ("From adafruit repo: allow
crystalless setup").  Code to use USBCRM was added back after having
been previously removed.  In the original Atmel code it was copied from,
the bitfield members of the register are used to set GCLK0's source as
OSC8M.  When this code was copied and converted to use the non-bitfield
access, the macro wasn't corrected.

The original Atmel code did this as an 8 MHz GCLK0 was used for the CPU
and UART.  Maybe in the "adafruit repo" this was copied from, an 8MHz
GCLK2 was used, or intended to be used, for something, but this was not
copied too, which is how the code to incorrectly configure an unused
clock was added.