Remove incorrect FDPLL96M 32kHz reference setting
Remove code to set the FDPLL 32 kHZ reference to GCLK0.  This saves 20 bytes of
code.

The FDPLL96M is not used in the bootloader.  The DFLL48M, which is used, is a
different clock.  The comment incorrectly states the DFLL reference clock is
being set, but that is ID 0, not ID 2.  Since USBCRM is used, the reference
clock for the DFLL is ignored and does not need to be set.

This code is setting a 32 kHz reference clock used for the FDPLL96M's internal
lock timer to GCLK0.  But GCLK0 is 48 MHz, so not only is this unnecessary, it's
also the wrong value.

The original Atmel bootloader had code to set the DFLL reference, but in that
bootloader GCLK0 was sourced from the internal OSC8M clock.  This code was
removed, but then added back.  When added back, the ID changed from 0 to 2, but
the comment wasn't changed.  Maybe someone wanted to set the DFLL ref to GCLK2?
Also unneeded, but would at least not be totally wrong.  This looks like some
cargo cult code that has just been carried along unnecessarily without doing
anything.